imports: fix vendored import paths
Before, vendored import paths would not be devendorized
for mocks generated using goautomock. This would fail
compilation for the generated code because vendored
import paths are not allowed to be used in Go's
import statements.

Now, the import paths are correctly devendorized by using
vendorlessImportPath() that was shamelessly copied form
https://github.com/golang/tools/blob/5cb1c80a83ef2a98bbc7b2fba1305239bb684fbe/imports/fix.go#L579-L590